refactor: decouple how to create formatted commits in tests
We specify a very particular `git log` format through a complex
`--pretty` format in `git-log.js`.

This means that if we do any change to the format, we have to edit
dozens of tests suites accordingly.

In order to mitigate this issue, we create a function that given a
commit subject and body generates output similar to out `--pretty`
format.

This way, if we ever change our format, we only have to update the tests
suite in a single place.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>