fix: parse commits with indented body
There are some commits we've seen in the wild while using this module
that indent the body for some reason (a space or two spaces usually).
For example:

```
My commit

 Lorem ipsum dolor sit amet.
 Hello world.

Fixes: #350
```

This confuses the YAML parser, since applying our custom `git log
--pretty` format to such commits yields:

```
- subject >-
    My commit
  body: |-
     Lorem ipsum dolor sit amet.
     Hello world.

    Fixes: #350
```

Which of course is not well indented.

We've considered moving away from YAML to another serialisation
technology that copes with these kind of multi-line string issues,
however we found a small workaround that allows us to not spend time
building and maintaing something else:

If we pass *any* dummy string as the first line of `body`, and make it
well indented, the remaining of the body will be parsed correctly.
Therefore, we tweak our parser to pass a dummy `XXX`, and implement code
in the parser to omit it.

Change-Type: patch
Changelog-Entry: Support commits with indented bodies.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>