fix: `stdout maxBuffer exceeded` when parsing big logs
`child_process.exec` imposes a limit on the largest amount of data
allowed on `stdout` or `stderr` (200 * 1024 bytes by default, according
to the documentation).

Since this limit can be easily reached when runnign Versionist on huge
projects, instead of executing the `git log` command, we spawn it,
and colelct the `stdout` output as it comes.

Change-Type: patch
Changelog-Entry: Fix `stdout maxBuffer exceeded` when parsing big logs.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>