fix: normalize template version
We were currently blindly passing the version the user sets as the
`--current` option to the Handlebars template, without imposing any
validation or normalisation.

This would cause problems with invalid semver versions, or if the user
sometimes passes a version preppended with a `v` and sometimes not,
since the data passed to the template would not always be consistent.

This PR introduces a function called `semver.normalize()`, which returns
the normalised version, or throws if the version is not valid.

Change-Type: patch
Changelog-Entry: Normalize `--current` version before passing it to the template.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>