feat: add `addEntryToChangelog` option hook
This hooks is used to determine how the CHANGELOG generated by
Versionist is added to the project's CHANGELOG file.

It ships with a `prepend` preset, which as the name implies, simply
prepends the entry to the CHANGELOG file specified with `changelogFile`,
which uses as a default value..

Change-Type: minor
Changelog-Entry: Add `addEntryToChangelog` option hook.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>