feat: add `getChangelogDocumentedVersions` option
This hook is used to determine which versions were already documented in
the project's CHANGELOG file specified in `changelogFile`.

The main purpose of this option is to prevent adding duplicate entries
to the CHANGELOG, and to calculate what's the latest documented version
automatically.

Change-Type: minor
Changelog-Entry: Add `getChangelogDocumentedVersions` option hook.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>