fix: make changelog entry addition idempotent
Currently, a generated entry is blindly added to the `CHANGELOG` file.
This means that if a user runs the same command over and over again, the
same duplicated entry will be added to the `CHANGELOG` multiple times.

This commit makes use of `getChangelogDocumentedVersions` to determine
whether the calculated version should be added to the CHANGELOG or not.

Change-Type: patch
Changelog-Entry: Make CHANGELOG entry addition idempotent.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>