fix: apply default values if the current value is not defined
There is a bug in the way we apply default values for boolean options.
If the user explicitly sets an option to `false`, then `property ||
default` will incorrectly be evaluated to `default`, which might be
`true.

Change-Type: patch
Changelog-Entry: Don't apply default values on boolean options explicitly set to `false`.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>