feat: add `editChangelog` option
This function makes Versionist add the generated entry to the CHANGELOG
automatically, which is the default behaviour.

Instead, if the option is set to `false`, the generated entry is simply
printed to `stdout`, which is very useful for debugging purposes, or
while the user is working on its template.

Change-Type: minor
Changelog-Entry: Add `editChangelog` option.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>