feat: infer `--from` and `--to`
Using Versionist is very manual. We require users to manually specify
the range of commits to be parsed with the `--from` and `--to` CLI
arguments.

Since the addition of the `getChangelogDocumentedVersions` hook, we can
use the information returned by that function to automatically infer
these values.

The CLI doesn't have the `--from`/`--to` options anymore and instead:

- Defaults `--to` to `HEAD`.
- Defaults `--from` to the greater version returned by
`getChangelogDocumentedVersions`.

Change-Type: major
Changelog-Entry: Automatically infer the appropriate git commit range.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>