docs: update README example to showcase new features
Recent commits introduced plenty features, like automatic `CHANGELOG`
and `package.json` editing.

The example from the `README.md` was updated to showcase this feature.

In order to make sure there are always enough changes to get something
generated, we added an `example` annotated tag and instruct the user to
`checkout` it.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>