feat: implement addEntryToChangelog `prepend` `fromLine` option
The CHANGELOG header was removed in 62c4da9 for simplicity purposes,
given we didn't have an easy way to deal with it on
`addEntryToChangelog` `prepend` preset.

This PR implements a `fromLine` option to the `prepend` preset, which as
the name implies, causes the prepending to happen from a certain line.

Now that we have this feature, we restore the original CHANGELOG header
and make use of `fromLine` in our `versionist.conf.js`.

Change-Type: minor
Changelog-Entry: Implement `fromLine` option for `addEntryToChangelog`'s `prepend` preset.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>