fix: uncaught exception if current version is omitted
The callbacks in this waterfall entry all yield a version. When omitting
the version, we call `callback` without parameters, therefore the next
waterfall entry passes the new callback as the first parameter, causing
a confusing error to be thrown.

In order to be consistent with other callback calls, we manually pass a
`null` version as the second argument.

Change-Type: patch
Changelog-Entry: Fix uncaught exception if current version is omitted.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>