feat: add `defaultInitialVersion` option
This option, which defaults to `0.0.1`, can be used to configure the
desired default initial version in case the
`getChangelogDocumentedVersions` hook doesn't find any version.

If the default version reference can't be found on git, then all commits
from the beginning of the project will be included in such version.

Change-Type: minor
Changelog-Entry: Add `defaultInitialVersion` option.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>