fix: reject non-existent git refs if there are documented versions
If we repeatedly call Versionist from the beginning of a project, all
the commits so far will be added over and over again.

The reason for this is that we default the start commit for
`versionist.readCommitHistory()` to `null` if a reference is not found,
independently on if it corresponds to the default initial version or
not.

To solve this issue, we only default the start commit to `null` if no
documented version was found on the repository.

Change-Type: patch
Changelog-Entry: Reject non-existent git references if there are documented versions.
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>