fix: ensure a specified config file is parsed correctly
Due to the way `yargs` operates, a named option cannot be used
as an object to pass back a new set of options (in this case
the loaded configuration), as the key on `argv` essentially already
exists.

To mitigate this, the returned object is renamed `configuration` and
this is used in place of `config` for all further operations.

Change-Type: minor
Changelog-Entry: Ensure a specified config file is parsed correctly.