Update the proper version definition of Rust Cargo manifests
We are doing more heavy regex processing now.

First the package name is extracted from Cargo.toml. It is located in the  `[package]` section of Cargo.toml. The package name is used for detecting the proper `version = "..."` definition in Cargo.lock. Cargo.lock is generated automatically, so `version` always follows `name`, and the code takes advantage of that fact for simplification purposes.

The proper `version = "..."` definition in Cargo.toml is inside the `[package]` section, so the regex looks for first occurrence after `[package]`.

Change-Type: patch
Connects-To: #101