Use smol-toml for parsing Cargo.toml
Replace regex usage with a more robust dependency.

This has the side effect of re-formatting `Cargo.toml` as this also uses
the `stringify` method of `smol-toml`, so removed a couple of tests that
checked for preservation of single quotes.

This also removes validation for Cargo.lock as that file should be
validated by cargo.

Change-type: patch